+++ cli.config.js
@@ -1,4 +1,0 @@
-import { getConfig } from '@shgysk8zer0/js-utils/rollup';
-import * as config from './config.mjs';
-
-export default getConfig('./cli.mjs', config);